requirements.txt
torch
torchvision
gradio
huggingface_hub
Pillow
numpy